fix(questions): add a property for the correct answer
This serves two purposes: the first is that questions that enable the
noShuffle option can place the correct answer in any position (and not
in the first which was the default).

The second purpose is that this fixes a bug where sometimes the correct
answer appeared as wrong. This happened because the answers were shuffled
without making a copy, which meant that the shuffling persisted. This
caused the first and "correct" answer to be different the second time
the same question appeared. Since making a copy of the array is not
convenient, since the whole object would need to be copied, and since
the order of the answers doesn't matter anyway, by storing the index of
the correct choice alongside with the array this problem is solved.